Move stuff from demo to core
